Extract Vault Directory to separate file
